Fix weird path bug with strtok
The man page for strtok says the string is separated by
"at least one of the characters in sep" but we're passing
a pointer to a character constant (which could be followed by
anything).

On my build (CLang, Mac OS X), PathSeparator was the '/' in
"/Animation". That gave some very weird results (trying to
create ..."/L/br/ry/C/ches/ge/rb").